MutationStorage integration test
